requirements.txt
mutagen>=1.47.0
python-dotenv>=1.0.1
pandas>=2.0.0
python-dateutil>=2.8.2
